requirements.txt
selenium>=4.15.0
webdriver-manager>=4.0.1
python-dotenv>=1.0.0
openai>=1.3.0
beautifulsoup4>=4.12.0
requests>=2.31.0
